Tests pour une extension des capacités du moteur, et essayer de contourner le segfault sur de très grand puzzles. Des tests sur une nouvelle version optimisée des chemins sont à venir.

+++ include/structs.h
@@ -6,7 +6,7 @@
 
 /*Structure d'une carte*/
 struct carte {
-  int identifiant;      /* L'identifiant de la carte */
+  long int identifiant;      /* L'identifiant de la carte */
   int Haut;             /* valeur se trouvant sur le haut de la carte */
   int Bas;              /* valeur du bas de la carte */
   int Gauche;           /* valeur de la gauche de la carte */
